Refactoring of copy, cut and paste commands with pattern Command.
