+++ src/drawingSoftware/Command.java
@@ -1,5 +1,8 @@
 package drawingSoftware;
 
+import drawingSoftware.Editor.Editor;
+
 public interface Command {
+    
     public void execute();
 }
